simplyfied code with hooks, IT WAS SOMETHING THAT I REALLY NEEDEd
